- Add contribution dashboard report

+++ database/migrations/2021_07_11_111751_create_plans_table.php
@@ -16,8 +16,8 @@ class CreatePlansTable extends Migration
         Schema::create('plans', function (Blueprint $table) {
             $table->id();
             $table->uuid('uuid')->unique();
-            $table->integer('amount');
-            $table->integer('fine_amount');
+            $table->decimal('amount', 10, 2);
+            $table->decimal('fine_amount', 10, 2);
             $table->timestamps();
         });
     }
