+++ database/migrations/2021_07_11_120701_create_payments_table.php
@@ -18,11 +18,11 @@ class CreatePaymentsTable extends Migration
             $table->uuid('uuid')->unique();
             $table->foreignId('patron_id')->constrained()->cascadeOnDelete();
             $table->foreignId('plan_id')->constrained();
-            $table->integer('amount');
-            $table->integer('fine')->default(0);
-            $table->integer('total');
-            $table->integer('paid')->default(0);
-            $table->integer('due');
+            $table->decimal('amount', 10, 2);
+            $table->decimal('fine', 10, 2)->default(0);
+            $table->decimal('total', 10, 2);
+            $table->decimal('paid', 10, 2)->default(0);
+            $table->decimal('due', 10, 2);
             $table->date('due_date');
             $table->unsignedTinyInteger('month')->default(1);
             $table->year('year');
